Initial addition of adding goals to the app.

+++ src/pages/home/home.ts
@@ -13,8 +13,6 @@ import { Subscription } from 'rxjs';
   templateUrl: 'home.html'
 })
 export class HomePage implements OnInit, OnDestroy {
-  //@ViewChild('delete_fab2') delete_fab;
-  //@ViewChild('edit_fab2') edit_fab;
 
   monthNames: string[] = ["January", "February", "March", "April", "May", "June",
   "July", "August", "September", "October", "November", "December"
@@ -29,6 +27,8 @@ export class HomePage implements OnInit, OnDestroy {
   dragging: boolean = false;
   fab_edit_display: boolean = false;
   fab_delete_display: boolean = false;
+  over_edit: boolean = false;
+  over_delete: boolean = false;
 
   constructor(public navCtrl: NavController,
               public modalCtrl: ModalController,
@@ -38,12 +38,14 @@ export class HomePage implements OnInit, OnDestroy {
   }
 
   ngOnInit() {
+
     this.subs.add(this.dragulaService.drag(this.dragulaName)
       .subscribe(({ name, el, source }) => {
         console.log("Drag event for task initiated.");
-
+        this.dragging = true;
         anime.remove("#delete_fab");
         anime.remove("#edit_fab");
+
         anime({
           targets: '#delete_fab',
           translateX: -100,
@@ -51,8 +53,6 @@ export class HomePage implements OnInit, OnDestroy {
           easing: 'easeOutElastic',
           elasticity: 300,
           duration: 500
-        }).finished.then(() => {
-          //this.fab_delete_display = true;
         });
 
         anime({
@@ -62,15 +62,15 @@ export class HomePage implements OnInit, OnDestroy {
           easing: 'easeOutElastic',
           elasticity: 300,
           duration: 500
-        }).finished.then(() => {
-          //this.fab_edit_display = true;
         });
+
       })
     );
 
     this.subs.add(this.dragulaService.dragend(this.dragulaName)
       .subscribe(({ name, el }) => {
         console.log("Drag ended.");
+        this.dragging = false;
         anime.remove("#delete_fab");
         anime.remove("#edit_fab");
         anime({
@@ -108,13 +108,14 @@ export class HomePage implements OnInit, OnDestroy {
 
           this.fab_edit_display = true;
           this.animateHover(1.2, 800, 400, "#edit_fab");
-
+          this.over_edit = true;
         }
         if(container.id == "delete_task") {
           console.log("Hovering over delete_task");
 
           this.fab_delete_display = true;
           this.animateHover(1.2, 800, 400, "#delete_fab");
+          this.over_delete = true;
         }
       })
     );
@@ -124,49 +125,45 @@ export class HomePage implements OnInit, OnDestroy {
         if(container.id == "edit_task") {
           console.log("Leaving edit_task");
           this.animateHover(1, 600, 300, "#edit_fab");
+          this.over_edit = false;
         }
         if(container.id == "delete_task") {
           console.log("Leaving delete_task");
           this.animateHover(1, 600, 300, "#delete_fab");
+          this.over_delete = false;
         }
       })
     );
 
-   this.subs.add(this.dragulaService.drop(this.dragulaName)
-    .subscribe(({ name, el, target, source, sibling }) => {
-      console.log("Drop event initiated.");
-      console.log("target: ");
-      console.log(target);
-      console.log("Source: ");
-      console.log(source);
-      console.log("Element: ");
-      console.log(el);
-      let taskIndex: number = Number(el.id);
-      if(target.id == "edit_task") {
-        this.editTask(this.sqliteService.tasks[taskIndex]);
-        this.dragulaService.find(this.dragulaName).drake.cancel(true);
-      }
-      else if(target.id == "delete_task") {
-        this.deleteTask(this.sqliteService.tasks[taskIndex].rowid);
-      }
-      else {
-        this.dragulaService.find(this.dragulaName).drake.cancel(true);
-      }
-
-    })
-  );
-
-
-
-    /*
-    this.subs.add(this.dragula.dragend(this.dragulaName)
-      .subscribe(({ name, el }) => {
-        console.log("Drag ended: ");
-        console.log("name: " + name + " element: ");
+    this.subs.add(this.dragulaService.drop(this.dragulaName)
+      .subscribe(({ name, el, target, source, sibling }) => {
+        /*
+        console.log("Drop event initiated.");
+        console.log("target: ");
+        console.log(target);
+        console.log("Source: ");
+        console.log(source);
+        console.log("Element: ");
         console.log(el);
+        console.log("Substring: ");
+        console.log(el.id.substring(5));
+        */
+        let taskRowId: number = Number(el.id.substring(5)); //draggable task id = "task_" + rowid, we just want the rowid.
+        let taskIndex: number = this.sqliteService.tasks.findIndex(item => item.rowid == taskRowId);
+
+        if(target.id == "edit_task" && this.over_edit) {
+          this.editTask(this.sqliteService.tasks[taskIndex]);
+          this.dragulaService.find(this.dragulaName).drake.cancel(true);
+        }
+        else if(target.id == "delete_task" && this.over_delete) {
+          this.deleteTask(this.sqliteService.tasks[taskIndex].rowid);
+        }
+        else {
+          this.dragulaService.find(this.dragulaName).drake.cancel(true);
+        }
+
       })
     );
-    */
 
   }
 
@@ -190,7 +187,7 @@ export class HomePage implements OnInit, OnDestroy {
     console.log('[home.ts] Entering ionViewDidLoad() --------------');
     this.todaysDate = new Date();
     this.date = new Date();
-    this.sqliteService.createAllTables();
+    //this.sqliteService.createAllTables();
     this.sqliteService.getTasksByDate(this.date).then((tasks: Task[]) => {
       this.sqliteService.tasks = tasks;
       this.sqliteService.sortTasks();
@@ -228,11 +225,13 @@ export class HomePage implements OnInit, OnDestroy {
   editTask(task: Task) {
     let presentEditTask = this.modalCtrl.create(EditTaskPage, {mode: "Edit", task: task});
 
+    /*
     presentEditTask.onDidDismiss(data => {
       if(data instanceof Task) {
 
       }
-    })
+    });
+    */
 
     presentEditTask.present();
   }
